♻️ : Move CommandLibrary into its own file
